// frontend/src/client/types.gen.ts
export type ContractCreate = {
    contract_number?: (string | null);
    contract_name?: (string | null);
    amount?: (number | null);
    sign_date?: (string | null);
    effective_date?: (string | null);
    expiry_date?: (string | null);
    project_id: string;
};